app/requirements.txt
requests == 2.32.4
urllib3 == 2.5.0
python-dotenv == 1.2.1
